+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Functions `getFilterState` to get current status of filters and `updateFilters` to update filters and refresh dataTable
+- Added function `getFilterState` to get current status of filters
+- Added function `updateFilters` to update filters and refresh dataTable
 
 ## [3.4.0] - 2023-10-03