+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.5.0] - 2023-10-05
+
 ### Added
 
 - Added function `getFilterState` to get current status of filters
@@ -181,7 +183,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - created package :tada:
 
-[unreleased]: https://github.com/neolution-ch/react-data-table/compare/3.4.0...HEAD
+[unreleased]: https://github.com/neolution-ch/react-data-table/compare/3.5.0...HEAD
+[3.5.0]: https://github.com/neolution-ch/react-data-table/compare/3.4.0...3.5.0
 [3.4.0]: https://github.com/neolution-ch/react-data-table/compare/3.3.0...3.4.0
 [3.3.0]: https://github.com/neolution-ch/react-data-table/compare/3.2.0...3.3.0
 [3.2.0]: https://github.com/neolution-ch/react-data-table/compare/3.1.0...3.2.0